Starting the refactor of the code using ReactJS
