+++ src/js/modelos.js
@@ -771,3 +771,5 @@ const db = {
         },
     ],
 }
+
+export default db
